Farther development of json game.

+++ api/ops/repair.php
@@ -410,7 +410,7 @@ class ApiPage extends OpsApiPageBase
 	function convert_op()
 	{
 		$last_id = 0;
-		$query = new DbQuery('SELECT id, log, end_time, canceled FROM games WHERE result > 0');
+		$query = new DbQuery('SELECT id, log, end_time, canceled, as_is FROM games WHERE result > 0');
 		if (isset($_REQUEST['last_id']))
 		{
 			$last_id = $_REQUEST['last_id'];
@@ -433,7 +433,7 @@ class ApiPage extends OpsApiPageBase
 		$c = 0;
 		foreach ($games as $row)
 		{
-			list($id, $log, $end_time, $is_canceled) = $row;
+			list($id, $log, $end_time, $is_canceled, $as_is) = $row;
 			$last_id = $id;
 			++$c;
 			try
@@ -455,22 +455,25 @@ class ApiPage extends OpsApiPageBase
 				$game->check(false);
 				
 				Db::begin();
-				Db::exec(get_label('game issue'), 'DELETE FROM game_issues WHERE game_id = ?', $id);
-				if (isset($game->issues))
+				if (!$as_is)
 				{
-					//echo 'Game ' . $id . "\n";
-					$old_json = $game->to_json();
-					$game->check($feature_flags, true);
-					if (isset($game->issues))
+					Db::exec(get_label('game issue'), 'DELETE FROM game_issues WHERE game_id = ?', $id);
+					if (!$as_is && isset($game->issues))
 					{
-						$issues = '<ul>';
-						foreach ($game->issues as $issue)
+						//echo 'Game ' . $id . "\n";
+						$old_json = $game->to_json();
+						$game->check($feature_flags, true);
+						if (isset($game->issues))
 						{
-							$issues .= '<li>' . $issue . '</li>';
+							$issues = '<ul>';
+							foreach ($game->issues as $issue)
+							{
+								$issues .= '<li>' . $issue . '</li>';
+							}
+							$issues .= '</ul>';
 						}
-						$issues .= '</ul>';
+						Db::exec(get_label('game issue'), 'INSERT INTO game_issues (game_id, json, issues) VALUES (?, ?, ?)', $id, $old_json, $issues);
 					}
-					Db::exec(get_label('game issue'), 'INSERT INTO game_issues (game_id, json, issues) VALUES (?, ?, ?)', $id, $old_json, $issues);
 				}
 				Db::exec(get_label('game'), 'UPDATE games SET json = ?, feature_flags = ? WHERE id = ?', $game->to_json(), $game->flags, $id);
 				Db::commit();
